Cleanup : deleted /include & renamed project
