+++ OpenGL/OpenGL/src/camera/Camera.cpp
@@ -1,5 +1,5 @@
 #include "Camera.hpp"
-#include "Shader.hpp"
+#include "../util/Shader.hpp"
 
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
