+++ OpenGL/OpenGL/src/controller/MouseController.cpp
@@ -1,4 +1,4 @@
-#include "controller/MouseController.hpp"
+#include "MouseController.hpp"
 
 #include <glm/gtc/matrix_transform.hpp>
 #include <iostream>
