+++ OpenGL/OpenGL/src/music/Music.cpp
@@ -1,5 +1,3 @@
-#include "include/Music.hpp"
-
 #include "Music.hpp"
 
 #include <iostream>
