+++ OpenGL/OpenGL/src/noise/bitmapTex.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "bitmap.hpp"
-#include "Shader.hpp"
+#include "util/Shader.hpp"
 
 /**
  * Draws a bitmap on screen. Mostly for debugging purposes.
