+++ OpenGL/OpenGL/src/terrain/Generator.hpp
@@ -2,7 +2,7 @@
 
 #include <glm/glm.hpp>
 #include "Chunk.hpp"
-#include "Noise.hpp"
+#include "noise/Noise.hpp"
 
 class Generator {
 public:
