+++ OpenGL/OpenGL/src/terrain/SkyBox.hpp
@@ -1,7 +1,7 @@
 #pragma once
-#include "../util/Loader.hpp"
-#include "Shader.hpp"
-#include "Camera.hpp"
+#include "util/Loader.hpp"
+#include "util/Shader.hpp"
+#include "camera/Camera.hpp"
 
 class SkyBox
 {
