+++ OpenGL/OpenGL/src/viewport/Viewport.hpp
@@ -2,9 +2,9 @@
 
 #include <string>
 
-#include "Camera.hpp"
-#include "controller/MouseController.hpp"
-#include "controller/KeyboardController.hpp"
+#include "../camera/Camera.hpp"
+#include "../controller/MouseController.hpp"
+#include "../controller/KeyboardController.hpp"
 #include "../entity/Hitbox.hpp"
 
 #include <SDL2/SDL.h>
